Lazy-load recharts in enrollment and grades popovers
Split chart rendering into dedicated modules loaded with next/dynamic
(ssr: false) so recharts stays out of the main schedule page bundle.
Reuse Skeleton for the brief chunk load on first popover open.
